backend/requirements.txt
fastapi
uvicorn
python-multipart
pdfplumber
diff-match-patch
reportlab
pypdfium2
Pillow
